Add workspace container
This container is intended to be used to actively develop the project, and
should provide all the tools needed for development. This includes a basic
text editor, version control, and all software needed to execute a build and
run the test suite.

It is not expected to actually run any of the project itself.